pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "astra-apod"
version = "0.1.0"
description = "A NASA Astronomy Picture of the Day CLI tool."
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
    "Pillow>=9.0.0",
    "numpy>=1.24.0",
]

[project.scripts]
astra = "astra.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["astra"]
